Extract mouse input main-thread cleanup
Move device release, delta reset, and timeout recovery behind one
stateless owner so action execution no longer carries cleanup policy.
Keep the async cancellation-token guard tracking the moved methods.
