+++ Packages/src/Editor/FirstPartyTools/SimulateMouseInput/MouseInputMainThreadCleanup.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7ab44557d0cc64ea1877f70eb1e036c1
+guid: eea8ef017defc4c7dbc892f72ddc6d8d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
